refactor(search applications): rename engine connect file
